Reference OCaml stdlib specifically
This bundles a special runtime module, Stdlib, which makes sure the
generated code uses the OCaml compiler runtime and not some code
outside it that just so happens to shadow the modules and functions
the code is using.

It also uses Base more consistently instead of using Base mixed
with the regular OCaml Stdlib.