Use the error monad inside the extension
Note that the extension code itself still raises an exception
if it detects an error in the [%mysql] statements (this is the
standard way for a PPX to signal an error).  However, with this
commit there is now a single place where that fatal exception
is raised.

Note also that the code generated by the PPX still relies
on exceptions internally (removing them is a separate issue).