Add support for cached statements
The idea is that prepared statements are cached for each DB connection.
Therefore the actual statement preparation only happens the first time
a given query is made for each connection.

This is implemented by wrapping the DB connection into a record that
also contains a hashtable with the statements already prepared. The
digest of the SQL query is used as the key to locate the statement.
(This is pretty much what PG'OCaml also does.)

Note that this required a breaking change in the API: user code must
now invoke Prepared.init to wrap a raw DB connection into a caching
connection.  Moreover, the PPX does not close statements anymore.