Replace `Pervasives` with `Stdlib`
Currently building on OCaml 4.08 fails with this error:

Error (alert deprecated): module Stdlib.Pervasives
Use Stdlib instead.

I don't quite get what the purpose of renaming that was and it breaks
all kinds of things now but that's how it is.

It also means that the newest supported version is 4.07 but oh well, I
guess that is fine.